[CBRD-23898] Fix errors detected by cppcheck (11.0) (#2752)
* [CBRD-23898] fix cppcheck errors other than broker and win_tools (#2673)

http://jira.cubrid.org/browse/CBRD-23898

Fix errors reported by cppcheck 2.4.1 in `src` source tree only.

Where in doubt, changes have been done conservatively as to not affect apparent existing behavior.

* [CBRD-23898] Fix for formatted print (#2722)

http://jira.cubrid.org/browse/CBRD-23898

The NULL checking for the buf should be performed only for unformatted print.

+++ src/base/tz_support.c
@@ -1892,7 +1892,9 @@ tz_encode_tz_id (const TZ_DECODE_INFO * tz_info, TZ_ID * tz_id)
 {
   if (tz_info->type == TZ_REGION_OFFSET)
     {
-      int offset = (tz_info->offset < 0) ? (-tz_info->offset) : tz_info->offset;
+      // *INDENT-OFF*
+      unsigned int offset = static_cast<unsigned int> ((tz_info->offset < 0) ? (-tz_info->offset) : tz_info->offset);
+      // *INDENT-ON*
 
       offset = offset & TZ_OFFSET_MASK;
